Downloads/finalappp/finalappp/requirements.txt
Pillow==8.4.0
numpy==1.21.2
matplotlib